msm_fb: fix color field offsets for 32-bit pixels
The offsets for pixel color fields were being reported precisely backwards. The
offsets are to be understood as right shifts within a 32-bit word. While the
bytes are ordered "R, G, B, A" in memory, they will appear in the order "ABGR"
when manipulated as a 32-bit word. Accordingly correct the offsets for formats
RGBA_8888 and ARGB_8888.

This matters when userspace is heeding the values reported by the framebuffer
driver, as does the Xorg X server. Android appears to ignore the values
and assume it knows the format of the pixels.