msm_vfe: don't assume ioctl arguments are valid
This ioctl handler immediately and directly dereferenced the arg
parameter without checking its validity or calling copy_from_user.
While this apparently worked with an Android userspace, which only
made valid ioctl calls, it causes oopses with other Linux userspaces
such as Pragmatux which call various ioctls to probe video devices for
identity and to determine their functionality.

Not using copy_from_user is, of course, never right, regardless of the
userspace.

This fix relies on the fact that there are no valid ioctls serviced
by this handler which leave arg null.

This patch doesn't completely repair the omission of copy_from_user calls
throught this driver or even in the patched function.

Signed-off-by: Ryan Kuester <rkuester@insymbols.com>