msm8960-perf_defconfig: configure out evbug driver
Don't build this driver, even as a module. If there's a udev around, it'll get
loaded in the presence of any event driver.  When loaded, its job is to print
every event to the kernel log.  In the presence of input devices like
accelerometers, which update continuously, this makes a real mess in a hurry.

It seems some distributions prevent this from loading with a modprobe
blacklist. That seems like a dumb way to "fix" the problem. The whole system
shouldn't be trying to load this in the first place.