Revert "msm_vfe: don't assume ioctl arguments are valid"
This reverts commit 67c33ae9b8f7ac9f68318144637e2bc358f8a95a.

This is an improper fix. Firstly, the arg doesn't need to be
copied_from_user because it is copied by video_usercopy early in the
call stack. Secondly, the real problem is that you can't assume what
command is reaching this fuction and what the contents of the arg
might be.