msm_vfe32: avoid OOPS when queried for capabilities
This is a quick hack to avoid OOPSing when queried for capabilities
(e.g., by udev via v4l_id). It seems this function naively assumes
that only a certain type of ioctl command will reach it; however, what
reaches it can be anything the core doesn't handle itself---which
could be almost anything.

The proper fix is probably to check the command against a whitelist,
but we'll leave that to the author. See other ioctl functions in the
same file. Note also that arg is dereferenced dangerously early,
because it could be null.